requirements.txt
finance-datareader
OpenDartReader
yfinance

# plot
matplotlib

# DB
psycopg2-binary

# ETC
requests
python-dotenv